Fixed all chargers not showing on map
Due to "bad" data from backend, where coordinates where integers instead of doubles, all of the chargers were not being fetched in the app. This problem is now solved and all of the chargers from the backend are being fetched despite having the wrong data type.

+++ lib/services/charger_api_service.dart
@@ -78,8 +78,8 @@ class ChargerApiService {
                       status: charger['status'],
                     )
                   ],
-                  coordinates:
-                      LatLng(charger['location'][0], charger['location'][1]),
+                  coordinates: LatLng(charger['location'][0].toDouble(),
+                      charger['location'][1].toDouble()),
                   price: chargerPoint.price,
                   name: chargerPoint.name,
                 ),
